requirements.txt
coverage==5.5
customtkinter==5.2.2
darkdetect==0.8.0
deprecation==2.1.0
eyed3==0.9.7
filetype==1.2.0
maturin==1.8.1
mutagen==1.47.0
packaging==24.2
patchelf==0.17.2.1
pillow==11.1.0
toml==0.10.2
